LPS-64610 Move
